+++ R/coefLognorm.R
@@ -1,99 +1,86 @@
+#' @describeIn getParmsLognormForModeAndUpper 
+#'   Calculates mu and sigma of lognormal from median and upper quantile.
+#' @param median geometric mu (median at the original exponential scale)
 #' @export
 getParmsLognormForMedianAndUpper <- function(
-  ### Calculates mu and sigma of lognormal from median and upper quantile.
-  median	##<< geometric mu (median at the original exponential scale)
-  ,quant	##<< value at the upper quantile, i.e. practical maximum
-  ,sigmaFac = qnorm(0.99) 	##<< sigmaFac=2 is 95% sigmaFac=2.6 is 99% interval
-){
+  median, upper,sigmaFac = qnorm(0.99)){
   # mu_geo = exp(mu); sigma_geo = exp(sigma)
   # logSpace cf-Interval: mu +- n sigma (n=1.96 for cf95)
   # expSpace cf-Interval: mu_geo */ simga_geo^n
   # given geometric mu (median at the original exponential scale)
   # and mu+sigmaFac*sigma
   tmp.mu = log(median)
-  cbind(mu = tmp.mu, sigma = 1/sigmaFac*(log(quant) - tmp.mu) )
-  ### named numeric vector: mu and sigma parameter of the lognormal distribution.
+  cbind(mu = tmp.mu, sigma = 1/sigmaFac*(log(upper) - tmp.mu) )
 }
 
+#' @describeIn getParmsLognormForModeAndUpper
+#'   Calculates mu and sigma of lognormal from mean and upper quantile.
+#' @param mean expected value at the original scale
+#' @details For \code{getParmsLognormForMeanAndUpper}
+#' there are two valid solutions, and the one with lower sigma
+#' , i.e. the not so strongly skewed solution is returned.
 #' @export
 getParmsLognormForMeanAndUpper <- function(
-  ### Calculates mu and sigma of lognormal from median and upper quantile.
-  mean	  ##<< expected value at the original scale
-  ,quant	##<< value at the upper quantile, i.e. practical maximum
-  ,sigmaFac = qnorm(0.99) 	##<< sigmaFac=2 is 95% sigmaFac=2.6 is 99% interval
+  mean, upper,sigmaFac = qnorm(0.99)
 ){
   # solution of
   # (1) mle = exp(mu - sigma^2)
-  # (2) quant = mu + sigmaFac sigma
+  # (2) upper = mu + sigmaFac sigma
   # see inst/doc/coefLognorm.Rmd for derivation
-  ##details<< There are two valid solutions. This routine returns
-  ## the one with lower sigma, i.e. the not so strongly skewed solution.
   m <- log(mean)
-  #sigma <- sigmaFac + sqrt(sigmaFac^2 - 2*(log(quant) - m))
-  sigma <- sigmaFac - sqrt(sigmaFac^2 - 2*(log(quant) - m))
+  #sigma <- sigmaFac + sqrt(sigmaFac^2 - 2*(log(upper) - m))
+  sigma <- sigmaFac - sqrt(sigmaFac^2 - 2*(log(upper) - m))
   mu <- m - sigma^2/2
   cbind(mu = mu, sigma = sigma )
-  ### numeric matrix: columns mu and sigma parameter of the
-  ### lognormal distribution.
 }
 
-
-
+#' @describeIn getParmsLognormForModeAndUpper
+#'   Calculates mu and sigma of lognormal from lower and upper quantile.
+#' @param lower value at the lower quantile, i.e. practical minimum
+#' @param upper value at the upper quantile, i.e. practical maximum
 #' @export
 getParmsLognormForLowerAndUpper <- function(
-  ### Calculates mu and sigma of lognormal from lower and upper quantile.
-  lower	  ##<< value at the lower quantile, i.e. practical minimum
-  , upper	##<< value at the upper quantile, i.e. practical maximum
-  , sigmaFac = qnorm(0.99) ##<< sigmaFac = 2 is 95%
-  ## sigmaFac = 2.6 is 99% interval
-  , isTransScale = FALSE ##<< if true lower and upper are already on log scale
-){
-  if (!isTRUE(isTransScale) ) {
-    lower <- log(lower)
-    upper <- log(upper)
-  }
-  sigma <- (upper - lower)/2/sigmaFac
-  cbind( mu = upper - sigmaFac*sigma, sigma = sigma )
-  ### named numeric vector: mu and sigma parameter of the lognormal
-  ### distribution.
-}
-attr(getParmsLognormForLowerAndUpper,"ex") <- function(){
-  # sample in normal space
-  mu <- 5
-  sigma <- 2
-  rrNorm <- rnorm(1000, mean = mu, sd = sigma)
-  # transform to orignal scale
-  rrOrig <- exp(rrNorm)
-  # and re-estimate parameters from original scale
-  res <- getParmsLognormForMedianAndUpper(
-    median(rrOrig), quantile(rrOrig, probs = 0.95), sigmaFac = qnorm(0.95))
-  expected <- c(mu = mu,sigma = sigma)
-  all.equal(res[1,], expected, tolerance = .1, scale = 1)
+  lower, upper, sigmaFac = qnorm(0.99)){
+  getParmsLognormForLowerAndUpperLog(log(lower), log(upper), sigmaFac)
 }
 
-
+#' @describeIn getParmsLognormForModeAndUpper
+#'   Calculates mu and sigma of lognormal from lower and upper quantile at log scale.
+#' @param lowerLog value at the lower quantile, i.e. practical minimum at log scale
+#' @param upperLog value at the upper quantile, i.e. practical maximum at log scale
 #' @export
+getParmsLognormForLowerAndUpperLog <- function(
+  lowerLog, upperLog, sigmaFac = qnorm(0.99)){
+  sigma <- (upperLog - lowerLog)/2/sigmaFac
+  cbind(mu = upperLog - sigmaFac*sigma, sigma = sigma)
+}
+
+#' Calculate mu and sigma of lognormal from summary statistics.
+#'
+#' @describeIn getParmsLognormForModeAndUpper 
+#'   Calculates mu and sigma of lognormal from mode and upper quantile.
+#' @param mle numeric vector: mode at the original scale
+#' @param upper numeric vector: value at the upper quantile,
+#'    i.e. practical maximum
+#' @param sigmaFac sigmaFac=2 is 95\% sigmaFac=2.6 is 99\% interval.
+#'
+#' @return numeric matrix wiht columns `mu` and `sigma`, the parameter of the
+#'   lognormal distribution. Rows correspond to rows of inputs.
+#' @export
+#' @exampleFunction example_getParmsLognormForModeAndUpper
 getParmsLognormForModeAndUpper <- function(
-  ### Calculates mu and sigma of lognormal from mode and upper quantile.
-  mle			  ##<< numeric vector: mode at the original scale
-  , quant		##<< numeric vector: value at the upper quantile,
-  ## i.e. practical maximum
-  , sigmaFac=qnorm(0.99) 	##<< sigmaFac=2 is 95% sigmaFac=2.6 is 99% interval
-){
+  mle, upper,sigmaFac = qnorm(0.99)){
   # solution of
   # (1) mle = exp(mu - sigma^2)
-  # (2) quant = mu + sigmaFac sigma
+  # (2) upper = mu + sigmaFac sigma
   # see inst/doc/coefLognorm.Rmd for derivation
   m <- log(mle)
   f2 <- sigmaFac/2
-  sigma <- -f2 + sqrt(f2^2 + (log(quant) - m))
+  sigma <- -f2 + sqrt(f2^2 + (log(upper) - m))
   mu <- m + sigma^2
   cbind(mu = mu, sigma = sigma )
-  ### numeric matrix: columns mu and sigma parameter of the
-  ### lognormal distribution.
-  ### Rows correspond to rows of mle and quant
 }
-attr(getParmsLognormForModeAndUpper,"ex") <- function(){
+example_getParmsLognormForModeAndUpper <- function(){
   # example 1: a distribution with mode 1 and upper bound 5
   (thetaEst <- getParmsLognormForModeAndUpper(1,5))
   mle <- exp(thetaEst[1] - thetaEst[2]^2)
@@ -109,8 +96,9 @@ attr(getParmsLognormForModeAndUpper,"ex") <- function(){
   theta0 <- c(mu = 1, sigma = 0.4)
   mle <- exp(theta0[1] - theta0[2]^2)
   perc <- 0.975		# some upper percentile, proxy for an upper bound
-  quant <- qlnorm(perc, meanlog = theta0[1], sdlog = theta0[2])
-  (thetaEst <- getParmsLognormForModeAndUpper(mle,quant = quant,sigmaFac = qnorm(perc)) )
+  upper <- qlnorm(perc, meanlog = theta0[1], sdlog = theta0[2])
+  (thetaEst <- getParmsLognormForModeAndUpper(
+    mle,upper = upper,sigmaFac = qnorm(perc)) )
 
   #plot the true and the rediscovered distributions
   xGrid = seq(0,10, length.out = 81)[-1]
@@ -132,17 +120,64 @@ attr(getParmsLognormForModeAndUpper,"ex") <- function(){
   matplot(x,dx,type = "l")
 }
 
-.ofLnormMLE <- function(
-  ### Objective function used by \code{\link{coefLogitnormMLE}}.
-  mu					  ##<< the mu parameter to estimate
-  ,logMle				##<< the mode of the density distribution
-  ,quant				##<< quantiles (values) at for percile perc
-  ,perc = 0.975 ##<< percentiles where quant is given
-){
-  #mle = exp( mu-sigma^2)
-  if (mu <= logMle) return(.Machine$double.xmax)
-  sigma = sqrt(mu - logMle)
-  tmp.predp = plnorm(quant, meanlog = mu, sdlog = sigma )
-  tmp.diff = tmp.predp - perc
-  sum(tmp.diff^2)
+#' @describeIn getParmsLognormForModeAndUpper 
+#'   Calculate mu and sigma from moments (mean anc variance)
+#' @param var variance at original scale
+#' @param sigmaOrig standard deviation at original scale 
+#' @references \code{Limpert E, Stahel W & Abbt M (2001)
+#' Log-normal Distributions across the Sciences: Keys and Clues.
+#' Oxford University Press (OUP) 51, 341,
+#' 10.1641/0006-3568(2001)051[0341:lndats]2.0.co;2}
+#' @export
+getParmsLognormForMoments <- function(mean, var, sigmaOrig = sqrt(var)){
+  omega = 1 + (sigmaOrig/mean)^2
+  mu = log(mean / sqrt(omega))
+  sigma = sqrt(log(omega))
+  cbind(mu,sigma)
 }
+
+
+#' @describeIn getParmsLognormForModeAndUpper 
+#'   Calculate mu and sigma from expected value and geometric standard deviation
+#' @param median geometric mu (median at the original exponential scale)
+#' @param sigmaStar multiplicative standard deviation
+#' @exampleFunction example_getParmsLognormForExpval
+#' @export
+getParmsLognormForExpval <- function(mean, sigmaStar){
+  sigma = log(sigmaStar)
+  mu = log(mean) - sigma*sigma/2
+  cbind(mu = mu, sigma = sigma)
+}
+example_getParmsLognormForExpval <- function(){
+  # Calculate mu and sigma from expected value and geometric standard deviation
+  .mean <- 1
+  .sigmaStar <- c(1.3,2)
+  (parms <- getParmsLognormForExpval(.mean, .sigmaStar))
+  # multiplicative standard deviation must equal the specified value
+  cbind(exp(parms[,"sigma"]), .sigmaStar)
+}
+
+#' Estimate lognormal distribution parameters from a sample
+#'
+#' @param x numeric vector of sampled values
+#' @param na.rm a logical value indicating whether 
+#' NA values should be stripped before the computation proceeds.
+#'
+#' @return numeric vector with components \code{mu} and \code{sigma},
+#' ie.., the center parameter (mean at log scale, log(median)) and 
+#' the scale parameter (standard deviation at log scale)
+#' 
+#' @export
+estimateParmsLognormFromSample <- function(x, na.rm = FALSE){
+  logx <- log(x)
+  c(mu = mean(logx, na.rm = na.rm) ,sigma = sd(logx, na.rm = na.rm))
+}
+example_estimateParmsLognormFromSample <- function(){
+  .mu <- log(1)
+  .sigma <- log(2)
+  x <- exp(rnorm(50, mean = .mu, sd = .sigma))
+  estimateParmsLognormFromSample(x)
+}
+
+
+