+++ R/imports.R
@@ -1,9 +1,9 @@
 #' @importFrom graphics abline axis box close.screen curve image legend 
-#'    lines mtext par plot points polygon screen split.screen
+#'    lines mtext par plot points polygon screen split.screen matplot
 #' @importFrom stats acf aggregate anova approx coef cor cov filter lm median 
 #'    na.omit na.pass nls nls.control optim predict quantile resid rnorm sd var 
 #'    AIC model.frame as.formula complete.cases
-#'    plnorm qnorm
+#'    plnorm qnorm dlnorm qlnorm
 NULL
 
 #' @importFrom Matrix Diagonal bandSparse